python/requirements.txt
opentelemetry-api
opentelemetry-sdk
boto3